criando arqivo das listas
