+++ index.html
@@ -6,6 +6,7 @@
   <title>*CCO INFORMA*</title>
   <link rel="shortcut icon" href="img/favicon.png" type="image/x-icon">
   <link rel="stylesheet" href="css/cco-informa.css">
+  <script nomodule src="js/lists.js" defer></script>
   <script nomodule src="js/data.js" defer></script>
   <script src="https://kit.fontawesome.com/f9a0d51bc2.js" crossorigin="anonymous" defer></script>
   <script type="module" src="js/creating-lists.js" defer></script>
